+++ server/server.js
@@ -6,20 +6,9 @@ const bodyParser = require('body-parser');
 const path = require('path');
 const csv = require('fast-csv');
 const dayjs = require('dayjs');
-const { validate: uuidValidate } = require('uuid');
-const multer = require('multer');
-const _ = require('lodash');
 const {
-  getParticipants,
   getHiredParticipantsBySite,
   getWithdrawnParticipantsBySite,
-  getParticipantByID,
-  updateParticipant,
-  parseAndSaveParticipants,
-  setParticipantStatus,
-  makeParticipant,
-  confirmParticipantInterest,
-  validateConfirmationId,
   getParticipantsForUser,
 } = require('./services/participants.js');
 const {
@@ -36,13 +25,8 @@ const {
   validate,
   EmployerFormSchema,
   AccessRequestApproval,
-  ParticipantQuerySchema,
-  ParticipantStatusChange,
-  ParticipantEditSchema,
-  ExternalHiredParticipantSchema,
   CreateSiteSchema,
   EditSiteSchema,
-  ParticipantSchema,
 } = require('./validation.js');
 const logger = require('./logger.js');
 const { dbClient, collections } = require('./db');
@@ -51,7 +35,13 @@ const keycloak = require('./keycloak.js');
 const { healthCheck } = require('./services/health-check');
 
 // Routes
-const participantUserRoute = require('./routes/participant-user');
+const {
+  participantUserRoute,
+  participantRouter,
+  participantsRouter,
+  newHiredParticipantRouter,
+  employerActionsRouter,
+} = require('./routes');
 
 const apiBaseUrl = '/api/v1';
 const app = express();
@@ -208,256 +198,6 @@ app.get(
   })
 );
 
-app.get(
-  `${apiBaseUrl}/participants/confirm-interest`,
-  asyncMiddleware(async (req, res) => {
-    const confirmationId = req.query.id;
-    const isValid = uuidValidate(confirmationId) && (await validateConfirmationId(confirmationId));
-    if (isValid) {
-      return res.status(200).send();
-    }
-
-    return res.status(400).send('Invalid Confirmation ID');
-  })
-);
-
-app.post(
-  `${apiBaseUrl}/participants/confirm-interest`,
-  asyncMiddleware(async (req, res) => {
-    const success = await confirmParticipantInterest(req.query.id);
-
-    if (success) {
-      return res.status(200).send();
-    }
-
-    return res.status(400).send('Invalid Confirmation ID');
-  })
-);
-
-// Get participant records
-app.get(
-  `${apiBaseUrl}/participants`,
-  keycloak.allowRolesMiddleware('health_authority', 'ministry_of_health', 'employer'),
-  keycloak.getUserInfoMiddleware(),
-  asyncMiddleware(async (req, res) => {
-    await validate(ParticipantQuerySchema, req.query);
-    const user = req.hcapUserInfo;
-    const {
-      offset,
-      regionFilter,
-      sortField,
-      sortDirection,
-      fsaFilter,
-      lastNameFilter,
-      emailFilter,
-      siteSelector,
-      statusFilters,
-    } = req.query;
-    const result = await getParticipants(
-      user,
-      {
-        pageSize: 10,
-        offset,
-        direction: sortDirection,
-      },
-      sortField,
-      regionFilter,
-      fsaFilter,
-      lastNameFilter,
-      emailFilter,
-      siteSelector,
-      statusFilters
-    );
-    logger.info({
-      action: 'participant_get',
-      performed_by: {
-        username: user.username,
-        id: user.id,
-      },
-      // Slicing to one page of results
-      ids_viewed: result.data.slice(0, 10).map((person) => person.id),
-    });
-    return res.json(result);
-  })
-);
-
-app.get(
-  `${apiBaseUrl}/participant`,
-  keycloak.allowRolesMiddleware('health_authority', 'ministry_of_health', 'employer'),
-  keycloak.getUserInfoMiddleware(),
-  asyncMiddleware(async (req, res) => {
-    await validate(ParticipantQuerySchema, req.query);
-    const user = req.hcapUserInfo;
-    const id = req.query;
-    const result = await getParticipantByID(id);
-    logger.info({
-      action: 'participant_get',
-      performed_by: {
-        username: user.username,
-        id: user.id,
-      },
-      on: {
-        id,
-      },
-    });
-    return res.json(result);
-  })
-);
-
-// Update participant data
-const patchableFields = [
-  'firstName',
-  'lastName',
-  'emailAddress',
-  'phoneNumber',
-  'interest',
-  'history',
-  'postalCode',
-  'postalCodeFsa',
-  'id',
-];
-
-app.patch(
-  `${apiBaseUrl}/participant`,
-  keycloak.allowRolesMiddleware('ministry_of_health'),
-  keycloak.getUserInfoMiddleware(),
-
-  asyncMiddleware(async (req, res) => {
-    req.body = Object.keys(req.body).reduce(
-      (o, k) => (patchableFields.includes(k) ? { ...o, [k]: req.body[k] } : o),
-      {}
-    );
-    await validate(ParticipantEditSchema, req.body);
-    const user = req.hcapUserInfo;
-    const result = await updateParticipant(req.body);
-    logger.info({
-      action: 'participant_patch',
-      performed_by: {
-        username: user.username,
-        id: user.id,
-      },
-    });
-    return res.json(result);
-  })
-);
-
-// Engage participant
-app.post(
-  `${apiBaseUrl}/employer-actions`,
-  keycloak.allowRolesMiddleware('health_authority', 'employer'),
-  keycloak.getUserInfoMiddleware(),
-  asyncMiddleware(async (req, res) => {
-    await validate(ParticipantStatusChange, req.body);
-    const user = req.hcapUserInfo;
-    const result = await setParticipantStatus(
-      user.id,
-      req.body.participantId,
-      req.body.status,
-      req.body.data
-    );
-    logger.info({
-      action: 'employer-actions_post',
-      performed_by: {
-        username: user.username,
-        id: user.id,
-      },
-      participant_id: req.body.participantId,
-      status: req.body.status,
-    });
-    return res.status(201).json({ data: result });
-  })
-);
-
-// Add Hired Participant to Database
-app.post(
-  `${apiBaseUrl}/new-hired-participant`,
-  keycloak.allowRolesMiddleware('employer', 'health_authority'),
-  keycloak.getUserInfoMiddleware(),
-  asyncMiddleware(async (req, res) => {
-    await validate(ExternalHiredParticipantSchema, req.body);
-    try {
-      const user = req.hcapUserInfo;
-      const participantInfo = req.body;
-      [participantInfo.preferredLocation] = user.regions;
-      participantInfo.crcClear = 'yes';
-      participantInfo.interested = 'yes';
-      participantInfo.callbackStatus = false;
-      participantInfo.userUpdatedAt = new Date().toJSON();
-
-      const response = await makeParticipant(participantInfo);
-      await setParticipantStatus(user.id, response.id, 'prospecting');
-      await setParticipantStatus(user.id, response.id, 'interviewing', {
-        contacted_at: participantInfo.contactedDate,
-      });
-      await setParticipantStatus(user.id, response.id, 'offer_made');
-      await setParticipantStatus(user.id, response.id, 'hired', {
-        site: participantInfo.site,
-        nonHcapOpportunity: !participantInfo.hcapOpportunity,
-        positionTitle: participantInfo.positionTitle,
-        positionType: participantInfo.positionType,
-        hiredDate: participantInfo.hiredDate,
-        startDate: participantInfo.startDate,
-      });
-
-      logger.info({
-        action: 'hired_participant_post',
-        performed_by: {
-          username: user.username,
-          id: user.id,
-        },
-        participant_id: response.id,
-      });
-
-      return res.status(201).json(response);
-    } catch (excp) {
-      return res.status(400).send(`${excp}`);
-    }
-  })
-);
-
-// Create participant records from uploaded XLSX file
-app.post(
-  `${apiBaseUrl}/participants/batch`,
-  keycloak.allowRolesMiddleware('maximus'),
-  keycloak.getUserInfoMiddleware(),
-  multer({
-    fileFilter: (req, file, cb) => {
-      if (file.fieldname !== 'file') {
-        req.fileError = 'Invalid field name.';
-        return cb(null, false);
-      }
-      if (file.mimetype === 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet') {
-        return cb(null, true);
-      }
-      req.fileError = 'File type not allowed.';
-      return cb(null, false);
-    },
-  }).single('file'),
-  asyncMiddleware(async (req, res) => {
-    if (req.fileError) {
-      return res.json({ status: 'Error', message: req.fileError });
-    }
-
-    try {
-      const response = await parseAndSaveParticipants(req.file.buffer);
-      const user = req.hcapUserInfo;
-      logger.info({
-        action: 'participant_post',
-        performed_by: {
-          username: user.username,
-          id: user.id,
-        },
-        // Slicing to one page of results
-        ids_posted: response.slice(0, 10).map((entry) => entry.id),
-      });
-
-      return res.status(201).json(response);
-    } catch (excp) {
-      return res.status(400).send(`${excp}`);
-    }
-  })
-);
-
 // Get linked participants for users
 app.get(
   `${apiBaseUrl}/user/participants`,
@@ -474,39 +214,6 @@ app.get(
       res.status(200).json(response);
     } else {
       res.status(401).send('Unauthorized user');
-    }
-  })
-);
-
-app.post(
-  `${apiBaseUrl}/participants`,
-  asyncMiddleware(async (req, res) => {
-    await validate(ParticipantSchema, req.body);
-
-    const participant = {
-      ...req.body,
-      formVersion: 'v2',
-      maximusId: null,
-      postalCodeFsa: req.body.postalCode.substr(0, 3),
-      preferredLocation: req.body.preferredLocation.join(';'),
-      interested: 'yes',
-      nonHCAP: null,
-      crcClear: null,
-      callbackStatus: false,
-      userUpdatedAt: new Date().toJSON(),
-    };
-
-    try {
-      const response = await makeParticipant(participant);
-      logger.info({
-        action: 'single_participant_post',
-        performed_by: null,
-        id: response.id,
-      });
-
-      return res.status(201).json(response);
-    } catch (excp) {
-      return res.status(500).send('Failed to create participant');
     }
   })
 );
@@ -816,6 +523,11 @@ app.get(
 
 // Applying router handlers
 app.use(`${apiBaseUrl}/participant-user`, participantUserRoute);
+
+app.use(`${apiBaseUrl}/participant`, participantRouter);
+app.use(`${apiBaseUrl}/participants`, participantsRouter);
+app.use(`${apiBaseUrl}/new-hired-participant`, newHiredParticipantRouter);
+app.use(`${apiBaseUrl}/employer-actions`, employerActionsRouter);
 
 // Client app
 if (process.env.NODE_ENV === 'production') {
